requirements.txt
flask
pymango
pymongo 
sentence-transformers 
requests
langchain-ollama
langchain_community
langchain_core
google-auth
google-auth-oauthlib
google-auth-httplib2
PyJWT
transformers
torch
torchvision
pillow